fix: correct hmr by not bundling server side app
- Server side application is not bundled. It is only transpiled by Babel. A main
reason for this is that when bundled, the application needs to know the required
modules at compile time. Though this could be configured with `externals` option
of webpack, clearing require.cache does not work as easily as expected.
Therefore, we determined to make server side app only transpiled with Babel,
and not bundled with Webpack.
- Module aliases configuration is moved to .babelrc.js. Module resolution is
done with `module-resolver`.`
- `webpack-merge` is removed from dependency.

Fixes https://github.com/eldeni/react-boilerplate/issues/24